services/insights/rag_pipeline/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
PyJWT>=2.10.0
aiomysql>=0.2.0
cachetools>=5.5.0
google-generativeai>=0.8.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
respx>=0.22.0
